Move get_and_set_proxy config function in pack_management module so it can be re-used by other code.
Also re-use existing do_register_cli_opts instead of copy and pasting it
over.
